new package structure for bdd
