+++ integration-tests/sts/sts_test.go
@@ -128,7 +128,7 @@ var _ = Describe("StatefulSet", func() {
 		// Check for pvc to get created and bound
 		Eventually(func() int {
 			pvcs, err := pvc.
-				NewKubeClient(pvc.WithNamespace(stsNamespace)).
+				NewKubeClient().WithNamespace(stsNamespace).
 				List(metav1.ListOptions{LabelSelector: stsApplicationLabel})
 			Expect(err).ShouldNot(HaveOccurred())
 			pvcCount, err := pvc.
@@ -160,7 +160,7 @@ var _ = Describe("StatefulSet", func() {
 		// Check for statefulset pods to get created and running
 		Eventually(func() int {
 			pods, err := pod.
-				NewKubeClient(pod.WithNamespace(stsNamespace)).
+				NewKubeClient().WithNamespace(stsNamespace).
 				List(metav1.ListOptions{LabelSelector: stsApplicationLabel})
 			Expect(err).ShouldNot(HaveOccurred())
 			return pod.
@@ -189,12 +189,12 @@ var _ = Describe("StatefulSet", func() {
 		stsApplicationLabel := "app=" + STSUnstructured.GetName()
 
 		// Fetch PVCs to be deleted
-		pvcs, err := pvc.NewKubeClient(pvc.WithNamespace(stsNamespace)).
+		pvcs, err := pvc.NewKubeClient().WithNamespace(stsNamespace).
 			List(metav1.ListOptions{LabelSelector: stsApplicationLabel})
 		Expect(err).ShouldNot(HaveOccurred())
 		// Delete PVCs
 		for _, p := range pvcs.Items {
-			err = pvc.NewKubeClient(pvc.WithNamespace(stsNamespace)).
+			err = pvc.NewKubeClient().WithNamespace(stsNamespace).
 				Delete(p.GetName(), &metav1.DeleteOptions{})
 			Expect(err).ShouldNot(HaveOccurred())
 		}
@@ -210,7 +210,7 @@ var _ = Describe("StatefulSet", func() {
 		// Verify deletion of sts instances
 		Eventually(func() int {
 			pods, err := pod.
-				NewKubeClient(pod.WithNamespace(stsNamespace)).
+				NewKubeClient().WithNamespace(stsNamespace).
 				List(metav1.ListOptions{LabelSelector: stsApplicationLabel})
 			Expect(err).ShouldNot(HaveOccurred())
 			return len(pods.Items)
@@ -220,7 +220,7 @@ var _ = Describe("StatefulSet", func() {
 		// Verify deletion of pvc instances
 		Eventually(func() int {
 			pvcs, err := pvc.
-				NewKubeClient(pvc.WithNamespace(stsNamespace)).
+				NewKubeClient().WithNamespace(stsNamespace).
 				List(metav1.ListOptions{LabelSelector: stsApplicationLabel})
 			Expect(err).ShouldNot(HaveOccurred())
 			return len(pvcs.Items)
@@ -252,7 +252,7 @@ var _ = Describe("StatefulSet", func() {
 			replicaAntiAffinityLabel := "openebs.io/replica-anti-affinity=" + STSUnstructured.GetName()
 
 			pvcs, err := pvc.
-				NewKubeClient(pvc.WithNamespace(stsNamespace)).
+				NewKubeClient().WithNamespace(stsNamespace).
 				List(metav1.ListOptions{LabelSelector: stsApplicationLabel})
 			Expect(err).ShouldNot(HaveOccurred())
 			pvcList, err := pvc.
